splitting the chart stuff into a dedicated pkg
